Add unit tests for WatchSpansAsync and WatchLogsAsync
Add 5 tests covering watcher functionality in TelemetryRepository:
- WatchSpansAsync_ReturnsExistingSpans_ThenNewSpans
- WatchSpansAsync_CanBeCancelled
- WatchSpansAsync_FiltersById_WhenResourceKeyProvided
- WatchLogsAsync_ReturnsExistingLogs_ThenNewLogs
- WatchLogsAsync_CanBeCancelled
